requirements.txt
async-timeout==4.0.2
boto3==1.28.1
botocore==1.31.1
certifi==2022.12.7
charset-normalizer==3.1.0
coverage==7.2.7
DateTime==5.1
distlib==0.3.6
exceptiongroup==1.1.1
filelock==3.12.2
greenlet==2.0.2
idna==3.4
iniconfig==2.0.0
jmespath==1.0.1
numpy==1.24.3
packaging==23.1
pandas==2.0.1
platformdirs==3.6.0
pluggy==1.0.0
pytest==7.3.2
pytest-cov==4.1.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
redis==4.6.0
requests==2.30.0
s3transfer==0.6.1
six==1.16.0
spotipy==2.23.0
SQLAlchemy==2.0.12
tomli==2.0.1
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.16
virtualenv==20.23.1
XlsxWriter==3.1.2
zope.interface==6.0
